refactor: announcement 분리 (#56)
