+++ src/main/java/com/example/eumserver/domain/announcement/resume/dto/ResumeAnnouncementRequest.java
@@ -1,4 +1,4 @@
-package com.example.eumserver.domain.resume.announcement.dto;
+package com.example.eumserver.domain.announcement.resume.dto;
 
 public record ResumeAnnouncementRequest(
         String introduction
