+++ src/main/java/com/example/eumserver/domain/announcement/resume/repository/ResumeAnnouncementRepository.java
@@ -1,6 +1,6 @@
-package com.example.eumserver.domain.resume.announcement.repository;
+package com.example.eumserver.domain.announcement.resume.repository;
 
-import com.example.eumserver.domain.resume.announcement.domain.ResumeAnnouncement;
+import com.example.eumserver.domain.announcement.resume.domain.ResumeAnnouncement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ResumeAnnouncementRepository extends JpaRepository<ResumeAnnouncement, Long> { }
