+++ src/main/java/com/example/eumserver/domain/announcement/team/controller/TeamAnnouncementController.java
@@ -1,12 +1,12 @@
-package com.example.eumserver.domain.team.announcement.controller;
+package com.example.eumserver.domain.announcement.team.controller;
 
-import com.example.eumserver.domain.team.announcement.domain.TeamAnnouncement;
-import com.example.eumserver.domain.team.announcement.dto.TeamAnnouncementFilter;
-import com.example.eumserver.domain.team.announcement.dto.TeamAnnouncementRequest;
-import com.example.eumserver.domain.team.announcement.dto.TeamAnnouncementResponse;
-import com.example.eumserver.domain.team.announcement.dto.TeamAnnouncementUpdateRequest;
-import com.example.eumserver.domain.team.announcement.mapper.TeamAnnouncementMapper;
-import com.example.eumserver.domain.team.announcement.service.TeamAnnouncementService;
+import com.example.eumserver.domain.announcement.team.domain.TeamAnnouncement;
+import com.example.eumserver.domain.announcement.team.dto.TeamAnnouncementFilter;
+import com.example.eumserver.domain.announcement.team.dto.TeamAnnouncementRequest;
+import com.example.eumserver.domain.announcement.team.dto.TeamAnnouncementResponse;
+import com.example.eumserver.domain.announcement.team.dto.TeamAnnouncementUpdateRequest;
+import com.example.eumserver.domain.announcement.team.mapper.TeamAnnouncementMapper;
+import com.example.eumserver.domain.announcement.team.service.TeamAnnouncementService;
 import com.example.eumserver.global.dto.ApiResult;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
