+++ src/main/java/com/example/eumserver/domain/announcement/team/domain/RecruitmentTypes.java
@@ -1,4 +1,4 @@
-package com.example.eumserver.domain.team.announcement.domain;
+package com.example.eumserver.domain.announcement.team.domain;
 
 /**
  * 모집 공고의 모집 유형을 나타낸 Enum Class 입니다.
