+++ src/main/java/com/example/eumserver/domain/announcement/team/dto/TeamAnnouncementFilter.java
@@ -1,6 +1,6 @@
-package com.example.eumserver.domain.team.announcement.dto;
+package com.example.eumserver.domain.announcement.team.dto;
 
-import com.example.eumserver.domain.team.announcement.domain.OccupationClassification;
+import com.example.eumserver.domain.announcement.team.domain.OccupationClassification;
 
 import java.util.List;
