+++ src/main/java/com/example/eumserver/domain/announcement/team/dto/TeamAnnouncementResponse.java
@@ -1,4 +1,4 @@
-package com.example.eumserver.domain.team.announcement.dto;
+package com.example.eumserver.domain.announcement.team.dto;
 
 public record TeamAnnouncementResponse(
         String title,
