+++ src/main/java/com/example/eumserver/domain/announcement/team/repository/TeamAnnouncementCustomRepositoryImpl.java
@@ -1,11 +1,11 @@
-package com.example.eumserver.domain.team.announcement.repository;
+package com.example.eumserver.domain.announcement.team.repository;
 
-import com.example.eumserver.domain.team.announcement.domain.OccupationClassification;
+import com.example.eumserver.domain.announcement.team.dto.TeamAnnouncementFilter;
+import com.example.eumserver.domain.announcement.team.dto.TeamAnnouncementResponse;
+import com.example.eumserver.domain.announcement.team.mapper.TeamAnnouncementMapper;
+import com.example.eumserver.domain.announcement.team.domain.OccupationClassification;
 import com.example.eumserver.domain.team.announcement.domain.QTeamAnnouncement;
-import com.example.eumserver.domain.team.announcement.dto.TeamAnnouncementFilter;
-import com.example.eumserver.domain.team.announcement.domain.TeamAnnouncement;
-import com.example.eumserver.domain.team.announcement.dto.TeamAnnouncementResponse;
-import com.example.eumserver.domain.team.announcement.mapper.TeamAnnouncementMapper;
+import com.example.eumserver.domain.announcement.team.domain.TeamAnnouncement;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
