+++ src/main/java/com/example/eumserver/domain/announcement/team/repository/TeamAnnouncementRepository.java
@@ -1,6 +1,6 @@
-package com.example.eumserver.domain.team.announcement.repository;
+package com.example.eumserver.domain.announcement.team.repository;
 
-import com.example.eumserver.domain.team.announcement.domain.TeamAnnouncement;
+import com.example.eumserver.domain.announcement.team.domain.TeamAnnouncement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TeamAnnouncementRepository extends JpaRepository<TeamAnnouncement, Long>, TeamAnnouncementCustomRepository {
