+++ src/main/java/com/example/eumserver/domain/resume/entity/Resume.java
@@ -1,6 +1,6 @@
 package com.example.eumserver.domain.resume.entity;
 
-import com.example.eumserver.domain.resume.announcement.domain.ResumeAnnouncement;
+import com.example.eumserver.domain.announcement.resume.domain.ResumeAnnouncement;
 import com.example.eumserver.domain.user.User;
 import com.example.eumserver.global.dto.TimeStamp;
 import com.fasterxml.jackson.annotation.JsonIgnore;
